Adding folders so that the files are more organized
