+++ app/src/main/java/com/example/fond/activities/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.fond;
+package com.example.fond.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.fond.R;
 
 public class MainActivity extends AppCompatActivity {
